Updated code for Czech keyboard
cz_CZ is syntactically incorrect. Code for Czech language is "cs", code for Czech Republic as a country is "CZ". So the common and correct code is "cs_CZ".

Also there are two commonly used keyboard layouts: "Czech", also called "Czech (QWERTZ)" (which is this layout) and "Czech (QWERTY)" which is currently not implemented. I'm working on it, but there are other changes than just Z/Y swap.